backend-part/requirements.txt
amqp==5.1.1
appdirs==1.4.4
asgiref==3.7.2
async-generator==1.10
attrs==23.1.0
beautifulsoup4==4.12.2
billiard==3.6.4.0
blinker==1.6.2
bs4==0.0.1
cachetools==5.3.1
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
et-xmlfile==1.1.0
exceptiongroup==1.1.1
Flask==2.3.2
Flask-Cors==3.0.10
google-api-core==2.11.0
google-api-python-client==2.88.0
google-auth==2.19.1
google-auth-httplib2==0.1.0
google-auth-oauthlib==1.0.0
googleapis-common-protos==1.59.0
h11==0.14.0
httplib2==0.22.0
idna==3.4
importlib-metadata==6.6.0
itsdangerous==2.1.2
Jinja2==3.1.2
kombu==5.3.0
MarkupSafe==2.1.2
mysql-connector-python==8.0.33
numpy==1.24.3
oauthlib==3.2.2
openpyxl==3.1.2
outcome==1.2.0
pandas==2.0.2
prompt-toolkit==3.0.38
protobuf==3.20.3
pyasn1==0.5.0
pyasn1-modules==0.3.0
pyee==8.2.2
pyparsing==3.0.9
pyppeteer==1.0.2
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2023.3
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
selenium==4.9.1
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.4.1
tqdm==4.65.0
trio==0.22.0
trio-websocket==0.10.2
typing_extensions==4.6.3
tzdata==2023.3
uritemplate==4.1.1
urllib3==1.26.16
vine==5.0.0
wcwidth==0.2.6
websockets==10.4
Werkzeug==2.3.4
wsproto==1.2.0
zipp==3.15.0
